Changed paths: M       CHANGES.txt M       cat_amap.mex M       cat_ornlm.mex M       cat_sanlm.mex M       cat_stat_check_cov.m M       cat_vbdist.mex M       cat_vol_downcut.c M       cat_vol_downcut.mex M       cat_vol_eidist.c M       cat_vol_eidist.mex M       cat_vol_eikonal3.c M       cat_vol_genus0.mex M       cat_vol_gradient3.c M       cat_vol_gradient3.mex M       cat_vol_interp3f.mex M       cat_vol_laplace3.c M       cat_vol_laplace3.mex M       cat_vol_laplace3R.c M       cat_vol_laplace3R.mex M       cat_vol_localstat.mex M       cat_vol_median3.mex M       cat_vol_median3c.mex M       cat_vol_pbtp.mex M       cat_vol_pbtp2.c M       cat_vol_simgrow.c M       cat_vol_simgrow.mex M       internal/cat_vol_cMRegularizarNLM3D.mex
(1) Changed: Move on with cov-check if report files are incomplete.
(2) Changed: Modified ind2sub subfunction for many c-files. 


git-svn-id: svn://141.35.69.214/cat12@1909 9f44b166-36c4-483c-8295-4af200601734

+++ cat_vol_eikonal3.c
@@ -25,9 +25,9 @@
 
 /* estimate x,y,z position of index i in an array size sx,sxy=sx*sy... */
 void ind2sub(int i,int *x,int *y, int *z, int sxy, int sy) {
-  *z = (int)floor( i / (double)sxy ) +1; 
+  *z = (int)floor( (double)i / (double)sxy ) +1; 
    i = i % (sxy);
-  *y = (int)floor( i / (double)sy ) +1;        
+  *y = (int)floor( (double)i / (double)sy ) +1;        
   *x = i % sy + 1;
 }
